% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comparison.R
\name{CompareKeyStats}
\alias{CompareKeyStats}
\title{Comparing key stats for stocks}
\usage{
CompareKeyStats(symbols, measures)
}
\arguments{
\item{symbols}{List of Yahoo Finance stock tickers}

\item{measures}{List of Key Stats to display. If empty display all that are available}
}
\description{
Comparing key stats for stocks
}
\examples{
CompareKeyStats(c("AAPL", "GOOG", "FB"), c("Operating Margin"))
}
\keyword{key}
\keyword{similar,}
\keyword{stats,}
\keyword{stocks}

