% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CorrelationMatrix.R
\name{CorrelationMatrix}
\alias{CorrelationMatrix}
\title{CorrelationMatrix.}
\usage{
CorrelationMatrix(ticker, cutoff = 0.8, start = Sys.Date() - 365,
  end = Sys.Date(), market.cap = 0.1, sector = TRUE, industry = FALSE)
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{start}{Start date}

\item{end}{End date}

\item{market.cap}{Percentage difference in market cap for similar stocks (or FALSE if not used)}

\item{sector}{Boolean to represent if stocks should be restricted to same sector}

\item{industry}{Boolean to represent if stocks should be restricted to same industry}

\item{correlation}{cutoff - used to determine if the other stock should be added to a basket of correlated stocks}
}
\value{
List of covariance matrix, correlation matrix, and basket of highly correlated stocks
}
\description{
CorrelationMatrix.
}
\examples{
CorrelationMatrix("IBM", 0.5, as.Date('2015-01-01'), as.Date('2015-05-25'))
}
\keyword{components}
\keyword{regression,}

