% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EarningsPerformance.R
\name{EarningsPerformance}
\alias{EarningsPerformance}
\title{Earnings Performance.}
\usage{
EarningsPerformance(ticker, searchStart = Sys.Date() - 365,
  searchEnd = Sys.Date(), timePeriod = 7)
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{searchStart}{Beginning of date range to search in}

\item{searchEnd}{End of date range to search in}

\item{timePeriod}{Period of time before and after earnings report}
}
\value{
data frame with earnings results
}
\description{
Earnings Performance.
}
\examples{
EarningsPerformance("TSLA", 2, 2015)
}
\keyword{earnings}
\keyword{performance}

