% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EventMovement.R
\name{EventMovement}
\alias{EventMovement}
\title{Event Performance}
\usage{
EventMovement(ticker, eventDates, timePeriod = 7)
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{eventDates}{list dates of events for which you want performance information}

\item{timePeriod}{period of time (in days) before and after events for which you want data}
}
\value{
vector of timeseries data showing performance in specified period before and after event dates
}
\description{
Event Performance
}
\examples{
EventMovement("TSLA", 2, 2015)
}
\keyword{earnings}
\keyword{performance}

