% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterComponents.R
\name{FilterComponents}
\alias{FilterComponents}
\title{Filter Components.}
\usage{
FilterComponents(ticker, components = "^GSPC", cutoff = 0.05,
  start = Sys.Date() - 365, end = Sys.Date())
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{components}{Vector of Yahoo Finance tickers to be filtered out}

\item{cutoff}{If p-value > cutoff, exclude the component}

\item{start}{Start date}

\item{end}{End date}
}
\value{
List of filtered out time series, returns, and regression results
}
\description{
Filter Components.
}
\examples{
FilterComponents("XOM", c("^GSPC", "USO"), 0.05, as.Date('2015-01-01'), as.Date('2015-05-25'))
}
\keyword{components}
\keyword{regression,}

