% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterTS.R
\name{FilterTS}
\alias{FilterTS}
\title{Filter Time Series}
\usage{
FilterTS(ts, components = "^GSPC", cutoff = 0.05)
}
\arguments{
\item{ts}{xts price series for a ticker}

\item{components}{Vector of Yahoo Finance tickers to be filtered out}

\item{cutoff}{If p-value > cutoff, exclude the component}
}
\value{
List of filtered out time series, returns, and regression results
}
\description{
Filter Time Series
}
\examples{
FilterTS(XOM, c("^GSPC", "USO"), 0.05)
}
\keyword{components}
\keyword{regression,}

