% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comparison.R
\name{GetKeyStats}
\alias{GetKeyStats}
\title{get key stats for a ticker}
\usage{
GetKeyStats(symbol)
}
\arguments{
\item{symbol}{Yahoo Finance stock ticker}
}
\description{
get key stats for a ticker
}
\examples{
CompareKeyStats(c("AAPL", "GOOG", "FB"), c("Operating Margin"))
}
\keyword{key}
\keyword{stats,}
\keyword{stocks}

