% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Comparison.R
\name{GetKeyStatsHelper}
\alias{GetKeyStatsHelper}
\title{used by GetKeyStats, credit to r-bloggers for this}
\usage{
GetKeyStatsHelper(symbol)
}
\arguments{
\item{symbol}{Yahoo Finance stock ticker}
}
\description{
used by GetKeyStats, credit to r-bloggers for this
}
\examples{
CompareKeyStats(c("AAPL", "GOOG", "FB"), c("Operating Margin"))
}
\keyword{key}
\keyword{stats,}
\keyword{stocks}

