% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HedgePortfolio.R
\name{HedgePortfolio}
\alias{HedgePortfolio}
\title{Hedge Portfolio}
\usage{
HedgePortfolio(portfolio, hedgePositions, cutoff = 0.05, start = Sys.Date()
  - 365, end = Sys.Date())
}
\arguments{
\item{portfolio}{character vector of Yahoo Finance tickers}

\item{hedgePositions}{1 : long on ticker, 0 : no exposure to ticker, -1 : short on ticker}

\item{cutoff}{If p-value > cutoff, exclude the portfolio component}

\item{start}{Start date}

\item{end}{End date}
}
\value{
numeric vector of hedge ratios for entire portfolio
}
\description{
Hedge Portfolio
}
\examples{
HedgePortfolio(c('SDRL', 'SDLP', 'SPY', 'USO'), c(-1, 1, 0, 0))
}
\keyword{components}
\keyword{regression,}

