% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HedgeTicker.R
\name{HedgeTicker}
\alias{HedgeTicker}
\title{Hedge Portfolio}
\usage{
HedgeTicker(ticker, numshares = 100, components = "^GSPC", cutoff = 0.05,
  start = Sys.Date() - 365, end = Sys.Date())
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{components}{Components that you want to have 0 exposure to when you hedge}

\item{cutoff}{If p-value > cutoff, exclude the portfolio component}

\item{start}{Start date}

\item{end}{End date}

\item{numShares}{number of shares in ticker}
}
\value{
number of shares of each component necessary to hedge out exposure
}
\description{
Hedge Portfolio
}
\examples{
HedgeTicker('SDRL', 100, c('SPY', 'USO'))
}
\keyword{components}
\keyword{regression,}

