% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LargeMoves.R
\name{LargeMoves}
\alias{LargeMoves}
\title{Large Moves.}
\usage{
LargeMoves(ticker, k, start = Sys.Date() - 365, end = Sys.Date())
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{k}{number of standard deviations to be considered a large move}

\item{start}{Start date}

\item{end}{End date}
}
\value{
Data frame of days that had large moves and their arithmetic returns
}
\description{
Large Moves.
}
\examples{
LargeMoves("XOM", 2, '2015-01-01', '2015-05-25')
}
\keyword{events}
\keyword{movement,}

