% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MovingCorrelation.R
\name{MovingCorrelation}
\alias{MovingCorrelation}
\title{MovingCorrelation}
\usage{
MovingCorrelation(ticker, symbols, start = Sys.Date() - 365,
  end = Sys.Date())
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{symbols}{vector of Yahoo Finance target tickets}

\item{start}{Start date}

\item{end}{End date}
}
\value{
lists of correlation coefficients
}
\description{
MovingCorrelation
}
\examples{
MovingCorrelation("SDRL", c("RIG","DO"),'2014-09-01', '2015-05-25')
}

