% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PairsBacktest.R
\name{PairsBacktest}
\alias{PairsBacktest}
\title{Pairs Backtest}
\usage{
PairsBacktest(ticker1, ticker2, capital = 1e+05, threshold = 1.5,
  startTrain = Sys.Date() - 365 * 2, endTrain = Sys.Date() - 365,
  startTest = Sys.Date() - 365, endTest = Sys.Date())
}
\arguments{
\item{ticker1}{First equity ticker}

\item{ticker2}{Second equity ticker}

\item{capital}{Starting cash}

\item{startTrain}{Beginning of correlation training period}

\item{endTrain}{End of correlation training period}

\item{startTest}{Beginning of backtest period}

\item{endTest}{End of backtest period}
}
\value{
vector of account values
}
\description{
Pairs Backtest
}
\keyword{arbitrage}
\keyword{backtest,}
\keyword{pairs,}
\keyword{statistical,}
\keyword{trade,}

