% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReturnMoments.R
\name{ReturnMoments}
\alias{ReturnMoments}
\title{Return Moments.}
\usage{
ReturnMoments(ticker, start = Sys.Date() - 365, end = Sys.Date())
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{start}{Start date}

\item{end}{End date}
}
\value{
vector of mean, SD, skewness, kurtosis
}
\description{
Return Moments.
}
\examples{
ReturnMoments("TSLA", '2015-01-01', '2015-05-25')
}
\keyword{moments}

