% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/SimilarStocks.R
\name{SimilarStocks}
\alias{SimilarStocks}
\title{Similar Stocks.}
\usage{
SimilarStocks(ticker, market.cap = 0.1, sector = TRUE, industry = FALSE)
}
\arguments{
\item{ticker}{Stock ticker}

\item{market.cap}{Percentage difference in market cap for similar stocks (or FALSE if not used)}

\item{sector}{Boolean to represent if stocks should be restricted to same sector}

\item{industry}{Boolean to represent if stocks should be restricted to same industry}
}
\value{
Data frame of similar stocks with their market caps, sectors, and industries
}
\description{
Similar Stocks.
}
\examples{
SimilarStocks("AAPL", market.cap = 0.15)
}
\keyword{cap,}
\keyword{industry}
\keyword{market}
\keyword{sector,}
\keyword{similar,}

