% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EarningsPerformance.R
\name{EarningsPerformance}
\alias{EarningsPerformance}
\title{Earnings Performance.}
\usage{
EarningsPerformance(ticker, quarterNumber, year)
}
\arguments{
\item{ticker}{Yahoo Finance target ticker}

\item{quarterNumber}{Integer from 1-4 representing Q1, Q2, etc.}

\item{year}{Year in which to search}
}
\value{
data frame with earnings results
}
\description{
Earnings Performance.
}
\examples{
EarningsPerformance("TSLA", 2, 2015)
}
\keyword{earnings}
\keyword{performance}

